changed the type of the capacity in stadium entity

+++ match-service/src/main/kotlin/com/youcode/kingsleague/match_service/models/dto/StadiumDTO.kt
@@ -6,10 +6,10 @@ data class StadiumDTO (
     var id: Long?,
     var name: String,
     var location: String,
-    var capacity: Int,
+    var capacity: Double,
     var matches: List<MatchDTO>?,
     var createdAt: LocalDateTime?,
     var updatedAt: LocalDateTime?
 ){
-    constructor() : this(null, "", "", 0, emptyList(), null, null)
+    constructor() : this(null, "", "", 0.00, emptyList(), null, null)
 }
